server/requirements.txt
amqp==5.0.6
arrow==1.1.1
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-celery-beat==2.2.1
django-filter==2.4.0
django-leaflet==0.28.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
django-cors-headers==3.8.0
dj_rest_auth==3.2.7
django-allauth==0.45.0
gunicorn==20.1.0
idna==3.2
kombu==5.1.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.3.2
polyline==1.4.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stravalib==0.10.4
units==0.7
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
